safari-534.53-branch/LayoutTests/svg/custom/selectSubString-expected.txt

PASS svgText.selectSubString(-1, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(0, -1) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS svgText.getSubStringLength(3, 0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS successfullyParsed is true

TEST COMPLETE
